requirements.txt
# engine
colorama==0.4.6
pysdl2==0.9.16
pysdl2-dll==2.28.5
python-ulid==2.4.0
Pillow==10.3.0
pyinstaller==6.5.0

# content build
bont
sprak
aseprite-reader
